Add support for custom domain name
Instead of metadatacenter.orgx one can now select a custom domain name,
which will be used as the $CEDAR_HOST env var. Beside this certificates
will be generated for this domain and the
keycloak-realm.CEDAR.development.20201020.json is edited to make keycloak
work with the custom domain.

Note: the ca.crt generated is still not perect. When imported into Firefox
it would still not make the cedar microservice domains valid for some
reason, so you still need to add each microservice url as a cert exception
by hand.